Allow consumers to disable installation
When including Vc as a subproject of another CMake project
(e.g. via FetchContent or git submodules) it might not
be necessary to not install Vc alongside the consuming project.

To allow consumers to decide whether they want to install Vc
or not this commit adds a new option Vc_ENABLE_INSTALL that
is ON by default.